添加低液位警报
1、添加警报声音、对话框
2、修正液位显示进度条Bug
3、CLevelMeter消息的m_LevelVal改为int类型，单位毫米

+++ LevelMeter.cpp
@@ -119,7 +119,7 @@ DWORD CLevelMeter::ProcMeter(SOCKET sockConn)
         goto RETURN;
     }
     UINT nLevelID = recvBuf[10] - '0';
-    float level;
+    int  level;
 
     for(; !m_nExit;)
     {
